starter-packs/gif-creator-pack/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gif-creator-pack"
version = "1.0.0"
description = "Create animated GIFs from images or text frames"
requires-python = ">=3.10"
license = "MIT"
dependencies = ["Pillow>=10.0"]

[tool.setuptools.packages.find]
where = ["src"]
